+++ bender.ci.js
@@ -3,7 +3,7 @@
 'use strict';
 var config = require( './bender' );
 
-config.startBrowser = 'Chrome';
+config.startBrowser = process.env.BROWSER || 'Chrome';
 config.mathJaxLibPath = 'https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS_HTML';
 
 module.exports = config;