+++ plugins/forms/dialogs/checkbox.js
@@ -116,10 +116,17 @@ CKEDITOR.dialog.add( 'checkbox', function( editor ) {
 						}
 					} else {
 						var value = this.getValue();
-						if ( value )
+						// Blink/Webkit needs to change checked property, not attribute. (#12465)
+						if ( CKEDITOR.env.webkit ) {
+							element.$.checked = value;
+						}
+
+						if ( value ) {
 							element.setAttribute( 'checked', 'checked' );
-						else
+						}
+						else {
 							element.removeAttribute( 'checked' );
+						}
 					}
 				}
 			},