+++ tests/core/selection/selection.js
@@ -652,5 +652,48 @@ bender.test( {
 		sel.getType();
 
 		assert.areSame( initialRev, sel.rev, 'Revision has not been modified' );
+	},
+
+	'test IE editable contenteditable="false" handling 1': function() {
+		if ( !CKEDITOR.env.ie ) {
+			assert.ignore();
+		}
+		var preventSpy = sinon.spy();
+
+		bender.tools.setHtmlWithSelection( this.editor, '<span contenteditable="false">^bar</span>' );
+
+		var selection = this.editor.getSelection(),
+			range = selection.getRanges()[ 0 ];
+
+		// Selection was moved somehow.
+		if ( range && range.startContainer.getName() !== 'span' ) {
+			assert.ignore();
+		}
+
+		this.editor.editable().fire( 'keydown', {
+			preventDefault: preventSpy,
+			getKeystroke: function() {},
+			getKey: function() {}
+		} );
+
+		assert.isTrue( preventSpy.called, 'preventDefault() on keydown was called' );
+	},
+
+	'test IE editable contenteditable="false" handling 2': function() {
+		if ( !CKEDITOR.env.ie ) {
+			assert.ignore();
+		}
+		var preventSpy = sinon.spy();
+
+		bender.tools.setHtmlWithSelection( this.editor, '<span contenteditable="false">' +
+			'<span contenteditable="true">^bar</span></span>' );
+
+		this.editor.editable().fire( 'keydown', {
+			preventDefault: preventSpy,
+			getKeystroke: function() {},
+			getKey: function() {}
+		} );
+
+		assert.isFalse( preventSpy.called, 'preventDefault() on keydown was called' );
 	}
 } );