+++ tests/core/style/applyremove.js
@@ -537,6 +537,21 @@
 				fixHtml( getInnerHtml( ct ).replace( /rgb\(255,255,0\)/g, 'rgb(255, 255, 0)' ) ) );
 		},
 
+		// #13062
+		'test forcing remove of boundary element': function() {
+			var editor = this.editor,
+				bot = this.editorBot,
+				style = new CKEDITOR.style( { element: 'b', type: CKEDITOR.STYLE_INLINE, alwaysRemoveElement: 1 } );
+
+			bot.setHtmlWithSelection( '<p><b>^example</b></p>' );
+			editor.removeStyle( style );
+			assert.areSame( '<p>^example</p>', bot.htmlWithSelection() );
+
+			bot.setHtmlWithSelection( '<p><b>example^</b></p>' );
+			editor.removeStyle( style );
+			assert.areSame( '<p>example^</p>', bot.htmlWithSelection() );
+		},
+
 		'test filler is preserved when applying block style': function() {
 			if ( !CKEDITOR.env.needsBrFiller )
 				assert.ignore();