+++ tests/plugins/link/link.js
@@ -475,13 +475,58 @@
 			} );
 		},
 
-		'test CKEDITOR.link.showDisplayTextForElement': function(){
+		'test CKEDITOR.link.showDisplayTextForElement': function() {
 			var doc = CKEDITOR.document,
 				showDisplayTextForElement = CKEDITOR.plugins.link.showDisplayTextForElement;
 
 			assert.isFalse( showDisplayTextForElement( doc.findOne( 'input#blurTarget' ), this.editor ), 'Input element' );
 			assert.isTrue( showDisplayTextForElement( doc.findOne( 'span' ), this.editor ), 'Span element' );
 			assert.isTrue( showDisplayTextForElement( null, this.editor ), 'Null value' );
+		},
+
+		// #13062
+		'test unlink when cursor is right before the link': function() {
+			var editor = this.editor,
+				bot = this.editorBot;
+
+			bot.setHtmlWithSelection( '<p><a href="http://cksource.com">^Link</a></p>' );
+
+			editor.ui.get( 'Unlink' ).click( editor );
+
+			assert.areSame( '<p>^Link</p>', bot.htmlWithSelection() );
+		},
+
+		// #13062
+		'test unlink when cursor is right after the link': function() {
+			// IE8 fails this test for unknown reason; however it does well
+			// in the manual one.
+			if ( CKEDITOR.env.ie && CKEDITOR.env.version == 8 ) {
+				assert.ignore();
+			}
+
+			var editor = this.editor,
+				bot = this.editorBot;
+
+			bot.setHtmlWithSelection( '<p><a href="http://cksource.com">Link^</a></p>' );
+
+			resume( function() {
+				editor.ui.get( 'Unlink' ).click( editor );
+				assert.areSame( '<p>Link^</p>', bot.htmlWithSelection() );
+			} );
+
+			wait( 100 );
+		},
+
+		// #13062
+		'test unlink when cursor is right before the link and there are more than one link in paragraph': function() {
+			var editor = this.editor,
+				bot = this.editorBot;
+
+			bot.setHtmlWithSelection( '<p>I am<a href="http://foo"> an </a>in<a href="http://bar">sta</a>nce of <a href="http://ckeditor.com">^<s>CKEditor</s></a>.</p>' );
+
+			editor.ui.get( 'Unlink' ).click( editor );
+
+			assert.areSame( '<p>I am<a href="http://foo"> an </a>in<a href="http://bar">sta</a>nce of ^<s>CKEditor</s>.</p>', bot.htmlWithSelection() );
 		}
 	} );
 } )();